test: rename file / class to follow convention
