+++ myapp/tests/test_BaseView.py
